refactor BPartnerComposite code
https://github.com/metasfresh/metasfresh/issues/5469
